Sama5d3: split install-sd into two pieces. Generate {uEnv,setenv}.txt from vars. Properly fetch zImage.

+++ files/sama5d3-xplained/prepare-sd.sh
@@ -6,14 +6,12 @@ set -e
 
 DISK="$1"
 
-#sudo dd if=/dev/zero of=${DISK} bs=1M count=16
-
 sudo sfdisk --in-order --Linux --unit M ${DISK} <<-__EOF__
 ,64,0xE,*
 ;
 __EOF__
 
-sleep 4
+sleep 3
 
 PART1="${DISK}1"
 PART2="${DISK}2"
@@ -21,6 +19,11 @@ if [ ! -b "${PART1}" ]
 then
 	PART1="${DISK}p1"
 	PART2="${DISK}p2"
+fi
+if [ ! -b "${PART1}" ]
+then
+	echo Partitions do not exist
+	exit 1
 fi
 
 sudo mkfs.vfat -F 16 "${PART1}" -n boot
